pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "kivy==2.3.1",
    "notebook>=7.5.3",
    "pillow==12.1.1",
    "pytest==9.0.2",
]
